:recycle: Adjust TransactionsOverviewHeader to be more similar to design
