+++ src/components/transactionsV2/transactionsOverviewHeader/transactionsOverviewHeader.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import { translate } from 'react-i18next';
 import { Link } from 'react-router-dom';
-import { getIndexOfFollowedAccount } from '../../../utils/followedAccounts';
 import { PrimaryButtonV2, SecondaryButtonV2 } from '../../toolbox/buttons/button';
 import RequestV2 from '../../requestV2/requestV2';
 import DropdownV2 from '../../toolbox/dropdownV2/dropdownV2';
-import AccountVisual from '../../accountVisual';
+import HeaderAccountInfo from './headerAccountInfo';
 import styles from './transactionsOverviewHeader.css';
 import routes from '../../../constants/routes';
 
@@ -44,49 +43,19 @@ class transactionsHeader extends React.Component {
 
   render() {
     const {
-      followedAccounts, address, t,
+      followedAccounts, address, t, delegate = {},
     } = this.props;
-    const index = getIndexOfFollowedAccount(
-      followedAccounts,
-      { address },
-    );
-    const accountTitle = index > -1 && followedAccounts[index]
-      && followedAccounts[index].title;
 
     const isMyWallet = address === this.props.account.address;
-    const hasTitle = !!accountTitle;
-    const placeholder = isMyWallet ? t('Wallet') : address;
 
     return (
       <header className={`${styles.wrapper}`}>
-        <div className={`${styles.account}`}>
-          <AccountVisual
-            address={address}
-            size={48}
-            />
-          <div className={styles.accountInfo}>
-            <div>
-              <h2 className={`${styles.title}`}>
-                <span className={'account-title'}>
-                { hasTitle ? accountTitle : placeholder }
-                </span>
-              </h2>
-              { isMyWallet ? (
-                <span className={`${styles.label} my-account`}>
-                  {t('My Account')}
-                </span>
-              ) : null }
-            </div>
-            { hasTitle || isMyWallet
-              ? (
-                <span className={styles.address}>
-                  {address}
-                </span>
-              ) : null
-            }
-          </div>
-        </div>
-
+        <HeaderAccountInfo
+          followedAccounts={followedAccounts}
+          address={address}
+          delegate={delegate}
+          account={this.props.account}
+          />
         <div className={`${styles.buttonsHolder}`}>
         { isMyWallet ?
           <span
